reorganization + add function
