+++ src/google_apis.py
@@ -15,7 +15,7 @@ def create_service(client_secret_file, api_name, api_version, *scopes, prefix=''
     SCOPES = [scope for scope in scopes[0]]
 
     cred = None
-    working_dir = os.getcwd()
+    working_dir = f"{os.getcwd()}/secrets"
     token_dir = 'token files'
     pickle_file = f'token_{API_SERVICE_NAME}_{API_VERSION}{prefix}.pickle'
