+++ src/main.py
@@ -1,5 +1,5 @@
 import time, random, os
-from datetime import datetime
+from datetime import datetime, timedelta
 from pathlib import Path
 import moviepy.editor as me
 
@@ -8,9 +8,9 @@ from src.dropbox import *
 from src.videos import *
 from src.youtube import *
 from src.helpers import *
-from src.helpers import *
 
-latest_video_title = get_latest_video_on_yt() # -> "The latest video on youtube"
+
+latest_video_title = get_latest_video_on_yt(os.getenv("YT_PLAYLIST_ID")) # -> "The latest video on youtube"
 
 podcasts_path = os.environ.get("PODCASTS_PATH") # -> "/Podcast"
 new_podcasts = get_new_podcasts(podcasts_path, latest_video_title)
@@ -28,7 +28,7 @@ print(f"{new_podcasts.lenght} new podcasts found...")
 last_youtube_publication_date = get_last_youtube_publication_date()
 for podcast in new_podcasts:
     podcast['publication_date'] = last_youtube_publication_date
-    last_youtube_publication_date = last_youtube_publication_date + datetime.timedelta(days=1)
+    last_youtube_publication_date = last_youtube_publication_date + timedelta(days=1)
     # @todo : set hour of publication_date to 6:00AM (with an env variable)
 
 for new_podcast in new_podcasts:
@@ -38,3 +38,4 @@ for new_podcast in new_podcasts:
     upload_video_to_youtube(video_path, thumbnail_path, new_podcast)
     delete_video_and_thumbnail(video_path, thumbnail_path)
     print(f"{new_podcast} processed")
+
